project2/requirements.txt
pyserial
matplotlib
numpy
sklearn
opencv-python
scikit-learn
pandas